Use attributeNAME!!!
For some reason, a Reactome object's list of attributes indicated that the "goBiologicalProcess" attribute originated in the "Reaction" class, when it really should originate in the "Event" class.

The attribute OBJECT for "goBiologicalProcess" is pulled from the attributes of an Event object so that attribute from the Event does not match the attribute with the same name in the Reaction because
they have different originating classes. Seems like it might be a bug in the data model itself, or the underlying code that loads objects from the database.

In the meantime, I'm changing the code to just use the attribute NAME. This will still get checked for validity, but it is a bit more tolerant of weirdness in the data model.

Also: Try printing the full stack trace to the log.